# api/openapi.yaml
openapi: 3.0.3
info:
  title: ForgeC2 Command & Control API
  description: |
    ForgeC2 is a professional-grade command and control framework for red team operations.

    ## Authentication
    Most endpoints require a session cookie (`forgec2_session`) from `POST /login`.
    Unauthenticated access to `/api/*` returns **JSON 401** (`{success:false,error:...}`), not HTML.

    ## CSRF
    State-changing requests need double-submit CSRF: cookie `forgec2_csrf` + header `X-CSRF-Token`.

    ## Spec quality (honest)
    - Route inventory is kept near-complete (CI coverage ≥90%, stale=0).
    - **Core operator paths** (health, agents, modules, listeners, dashboard, deploy)
      include request/response schemas; many secondary routes remain summary-only.
    - Dual-use paths (e.g. `GET /agents`) serve SPA HTML or JSON depending on Accept/auth.
    - `/debug/pprof/*` is intentionally omitted (disabled by default).

    ## Rate Limiting
    - Login: Limited attempts with lockout
    - API: Token bucket rate limiting per user

    ## Permissions
    Role-based access control (custom roles supported):
    - **admin**: Full control (all permissions)
    - **user**: Standard operator permissions (agents, tasks, credentials, limited settings)

  version: 2.4.1
  contact:
    name: ForgeC2 Support
    email: support@forgec2.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
...
  - url: https://c2.example.com
...
paths:
...
  /api/agents:
    get:
      tags: [Agents]
      summary: List all agents
      description: Retrieve a paginated list of all agents with optional filters
      operationId: listAgents
      security:
        - SessionCookie: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Search by hostname, username, or IP
        - name: status
          in: query
          schema:
            type: string
            enum: [online, stale, offline]
          description: Filter by agent status
        - name: os
          in: query
          schema:
            type: string
          description: Filter by operating system
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Page number
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
          description: Items per page
        - name: sort_key
          in: query
          schema:
            type: string
            enum: [hostname, username, os, ip, last_seen, version, status]
            default: last_seen
          description: Sort by column (status orders by derived online/stale/offline)
        - name: sort_dir
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort direction
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Agent'
                  total:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /agents/{id}:
    get:
      tags: [Agents]
      summary: Get agent detail page
      description: Render agent detail HTML page
      operationId: getAgentDetailPage
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Agent detail page
          content:
            text/html:
              schema:
                type: string
        '404':
          description: Agent not found

    delete:
      tags: [Agents]
      summary: Delete agent
      description: Remove an agent from the system
      operationId: deleteAgent
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Agent deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /agents/{id}/tasks:
    get:
      tags: [Tasks, Agents]
      summary: List agent tasks
      description: Retrieve all tasks for a specific agent
      operationId: listAgentTasks
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, running, completed, failed]
          description: Filter by task status
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /agents/{id}/tasks/{taskId}:
    get:
      tags: [Tasks, Agents]
      summary: Get task status
      description: Retrieve detailed information about a specific task
      operationId: getTaskStatus
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/listeners:
    get:
      tags: [Listeners]
      summary: List all listeners
      description: Retrieve all configured listeners with pagination and filtering
      operationId: listListeners
      security:
        - SessionCookie: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Page number
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
          description: Items per page
        - name: tag
          in: query
          schema:
            type: string
          description: Filter by tag (substring match)
        - name: status
          in: query
          schema:
            type: string
            enum: [running, stopped, error]
          description: Filter by status
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Listener'
                  total:
                    type: integer
                  page:
                    type: integer
                  page_size:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Listeners]
      summary: Create listener
      description: Create a new listener configuration
      operationId: createListener
      security:
        - SessionCookie: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListenerCreate'
      responses:
        '201':
          description: Listener created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Listener'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /credentials:
    get:
      tags: [Credentials]
      summary: Credentials page
      description: Render credentials vault HTML page
      operationId: getCredentialsPage
      security:
        - SessionCookie: []
      responses:
        '200':
          description: Credentials page
          content:
            text/html:
              schema:
                type: string
...
  /settings/server:
    post:
      tags: [Settings]
      summary: Save server config
      description: Save server configuration
      operationId: saveServerConfig
      security:
        - SessionCookie: []
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                log_level:
                  type: string
                  enum: [debug, info, warn, error]
                tcp_enabled:
                  type: boolean
                tcp_addr:
                  type: string
                offline_threshold:
                  type: integer
                session_max_age:
                  type: integer
                cleanup_retention:
                  type: integer
      responses:
        '200':
          description: Config saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
  /audit:
    get:
      tags: [Audit]
      summary: Audit log page
      description: Render audit log HTML page
      operationId: getAuditLogPage
      security:
        - SessionCookie: []
      responses:
        '200':
          description: Audit log page
          content:
            text/html:
              schema:
                type: string
...
  /tasks:
    get:
      tags: [Tasks]
      summary: Task history page
      description: Render task history HTML page
      operationId: getTaskHistory
      security:
        - SessionCookie: []
      responses:
        '200':
          description: Task history page
          content:
            text/html:
              schema:
                type: string
...
  /tasks/{taskId}:
    get:
      tags: [Tasks]
      summary: Get task details
      description: Get detailed information about a task
      operationId: getTaskDetails
      security:
        - SessionCookie: []
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /agents/bulk/results:
    get:
      tags: [Agents]
      summary: Bulk task results
      operationId: agentsBulkResults
      security:
        - SessionCookie: []
      responses:
        '200':
          description: OK
...
  /api/v1/agents/{id}:
    delete:
      tags: [API]
      summary: DELETE /api/v1/agents/{id}
      operationId: delete_api_v1_agents_id
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
    get:
      tags: [API]
      summary: GET /api/v1/agents/{id}
      operationId: get_api_v1_agents_id
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
...
  /api/v1/audit:
    get:
      tags: [API]
      summary: GET /api/v1/audit
      operationId: get_api_v1_audit
      security:
        - SessionCookie: []
      responses:
        '200':
          description: OK
  /api/v1/bulk/results:
    get:
      tags: [API]
      summary: GET /api/v1/bulk/results
      operationId: get_api_v1_bulk_results
      security:
        - SessionCookie: []
      responses:
        '200':
          description: OK
...
  /api/v1/task-types:
    get:
      tags: [API]
      summary: GET /api/v1/task-types
      operationId: get_api_v1_task_types
      security:
        - SessionCookie: []
      responses:
        '200':
          description: OK
  /api/v1/tasks/{id}:
    get:
      tags: [API]
      summary: GET /api/v1/tasks/{id}
      operationId: get_api_v1_tasks_id
      security:
        - SessionCookie: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: OK
...
  /api/v1/tasks/status:
    post:
      tags: [API]
      summary: POST /api/v1/tasks/status
      operationId: post_api_v1_tasks_status
      security:
        - SessionCookie: []
      responses:
        '200':
          description: OK
...
  /api/v1/dashboard:
    get:
      tags: [Agents]
      summary: Dashboard stats (REST v1)
      operationId: apiV1Dashboard
      security:
        - SessionCookie: []
      responses:
        '200':
          description: Dashboard payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardStats'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/agents:
    get:
      tags: [Agents]
      summary: List agents (REST v1)
      operationId: apiV1ListAgents
      security:
        - SessionCookie: []
      responses:
        '200':
          description: Agent list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/tasks:
    get:
      tags: [Tasks]
      summary: List tasks (REST v1)
      operationId: apiV1ListTasks
      security:
        - SessionCookie: []
      responses:
        '200':
          description: Task list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  tasks:
                    type: array
                    items:
                      $ref: '#/components/schemas/Task'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Tasks]
      summary: Create task (REST v1)
      operationId: apiV1CreateTask
      security:
        - SessionCookie: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                agent_id: { type: string }
                type: { type: string }
                command: { type: string }
      responses:
        '200':
          description: Task created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
...
  /agents:
    get:
      tags: [Agents]
      summary: List agents (dual-use SPA/API path)
      description: |
        With `Accept: application/json` returns agent list JSON.
        Browser navigation may receive SPA HTML. Unauthenticated JSON requests return 401.
      operationId: listAgentsLegacy
      security:
        - SessionCookie: []
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema:
            type: string
            enum: [online, stale, offline]
      responses:
        '200':
          description: Agent list (JSON) or SPA shell
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentListResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/listeners:
    get:
      tags: [Listeners]
      summary: List listeners (REST v1)
      operationId: apiV1ListListeners
      security:
        - SessionCookie: []
      responses:
        '200':
          description: Listeners
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Listener'
  /api/v1/credentials:
    get:
      tags: [Credentials]
      summary: List credentials (REST v1)
      operationId: apiV1ListCredentials
      security:
        - SessionCookie: []
      responses:
        '200':
          description: Credentials
...
  /listeners:
    get:
      tags: [Listeners]
      summary: Listeners page list
      operationId: get__listeners
      security:
        - SessionCookie: []
      responses:
        '200':
          description: OK
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: forgec2_session
      description: Session cookie obtained from login endpoint
  schemas:
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
    TaskResponse:
      type: object
      required: [success, task_id]
      properties:
        success:
          type: boolean
          example: true
        task_id:
          type: integer
    Error:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        code:
          type: string
        details:
          type: string
...
    DashboardStats:
      type: object
      description: Aggregate agent/task counters for the operator dashboard
      properties:
        success:
          type: boolean
        total_agents:
          type: integer
        online_agents:
          type: integer
        stale_agents:
          type: integer
        offline_agents:
          type: integer
        total_tasks:
          type: integer
        tasks_today:
          type: integer
        pending_tasks:
          type: integer
        failed_tasks:
          type: integer
        total_listeners:
          type: integer
...
    Agent:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique agent identifier
        hostname:
          type: string
        username:
          type: string
        os:
          type: string
          example: Windows 10 Pro
        arch:
          type: string
          example: amd64
        ip:
          type: string
        public_ip:
          type: string
        country:
          type: string
        city:
          type: string
        last_seen:
          type: string
          format: date-time
        status:
          type: string
          enum: [online, stale, offline]
        notes:
          type: string
        tags:
          type: string
        listener_id:
          type: integer
        parent_id:
          type: string
          description: Parent agent ID for P2P chaining
        p2p_mode:
          type: string
          enum: [smb, tcp, ""]
        p2p_listen_addr:
          type: string
        version:
          type: string
        pid:
          type: integer
        process_name:
          type: string
        integrity:
          type: string
          enum: [Low, Medium, High, System]
        elevated:
          type: boolean
        domain:
          type: string
        current_interval:
          type: integer
          description: Sleep interval in seconds
        current_jitter:
          type: integer
          description: Jitter percentage
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Task:
      type: object
      properties:
        id:
          type: integer
        agent_id:
          type: string
          format: uuid
        type:
          type: string
          description: Task type
        command:
          type: string
          description: Primary command or path
        shell:
          type: string
          description: Shell choice or secondary data
        path:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed]
        result:
          type: string
        error:
          type: string
        progress:
          type: integer
          description: 0-100 percentage
        total_bytes:
          type: integer
          format: int64
        transferred:
          type: integer
          format: int64
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Listener:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        scheme:
          type: string
          enum: [http, https, tcp, tls, dns, icmp, ssh, h2c]
        type:
          type: string
          enum: [http, tcp, dns, icmp, ssh, h2c]
        host:
          type: string
        port:
          type: integer
        protocol:
          type: string
        notes:
          type: string
        enabled:
          type: boolean
        tags:
          type: string
          description: Comma-separated tags for categorization
        color:
          type: string
          description: Hex color code for UI display
        status:
          type: string
          enum: [running, stopped, error]
          description: Runtime status
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ListenerCreate:
      type: object
      required: [name, scheme, host, port]
      properties:
        name:
          type: string
        scheme:
          type: string
          enum: [http, https, tcp, tls, dns, icmp, ssh, h2c]
        host:
          type: string
        port:
          type: integer
        notes:
          type: string
        tags:
          type: string
          description: Comma-separated tags for categorization
        color:
          type: string
          description: Hex color code for UI display
        status:
          type: string
          enum: [running, stopped, error]
          description: Runtime status
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - not authenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'